Added menus and keyboard shortcuts
